+++ WultraMobileTokenSDKTests/IntegrationProxy.swift
@@ -23,6 +23,7 @@ class IntegrationProxy {
     private(set) var powerAuth: PowerAuthSDK?
     private(set) var operations: WMTOperations?
     private(set) var inbox: WMTInbox?
+    private(set) var push: WMTPush?
     
     private var config: IntegrationConfig!
     private let activationName = UUID().uuidString
@@ -30,10 +31,11 @@ class IntegrationProxy {
     
     typealias Callback = (_ error: String?) -> Void
     
-    func prepareActivation(pin: String, callback: @escaping Callback) {
+    func prepareActivation(pin: String, configFileName: String = "config", callback: @escaping Callback) {
         WPNLogger.verboseLevel = .debug
-        guard let configPath = Bundle.init(for: IntegrationProxy.self).path(forResource: "config", ofType: "json", inDirectory: "Configs") else {
-            callback("Config file config.json is not present.")
+        
+        guard let configPath = Bundle.init(for: IntegrationProxy.self).path(forResource: configFileName, ofType: "json", inDirectory: "Configs") else {
+            callback("Config file \(configFileName).json is not present.")
             return
         }
         
@@ -41,7 +43,7 @@ class IntegrationProxy {
             let configContent = try String(contentsOfFile: configPath)
             config = try JSONDecoder().decode(IntegrationConfig.self, from: configContent.data(using: .utf8)!)
         } catch _ {
-            callback("Config file config.json cannot be parsed.")
+            callback("Config file \(configFileName).json cannot be parsed.")
             return
         }
         
@@ -52,9 +54,11 @@ class IntegrationProxy {
             } else {
                 let wpnOperationsConf = WPNConfig(baseUrl: URL(string: self.config.operationsServerUrl)!, sslValidation: .noValidation)
                 let wpnInboxConf = WPNConfig(baseUrl: URL(string: self.config.inboxServerUrl)!, sslValidation: .noValidation)
+                let wpnPushConf = WPNConfig(baseUrl: URL(string: self.config.pushServerUrl)!, sslValidation: .noValidation)
                 self.powerAuth = pa
                 self.operations = pa.createWMTOperations(networkingConfig: wpnOperationsConf, pollingOptions: [.pauseWhenOnBackground])
                 self.inbox = pa.createWMTInbox(networkingConfig: wpnInboxConf)
+                self.push = pa.createWMTPush(networkingConfig: wpnPushConf)
                 callback(nil)
             }
         }
@@ -302,6 +306,7 @@ private struct IntegrationConfig: Codable {
     let enrollmentServerUrl: String
     let operationsServerUrl: String
     let inboxServerUrl: String
+    let pushServerUrl: String
     let sdkConfig: String
 }